HSEARCH-5133 Push factories into each Lucene*Aggregation
