+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/types/aggregation/impl/AbstractLuceneMetricNumericFieldAggregation.java
@@ -12,7 +12,6 @@ import org.hibernate.search.backend.lucene.lowlevel.collector.impl.CollectorKey;
 import org.hibernate.search.backend.lucene.lowlevel.docvalues.impl.JoiningLongMultiValuesSource;
 import org.hibernate.search.backend.lucene.search.aggregation.impl.AggregationExtractContext;
 import org.hibernate.search.backend.lucene.search.aggregation.impl.AggregationRequestContext;
-import org.hibernate.search.backend.lucene.search.common.impl.AbstractLuceneCodecAwareSearchQueryElementFactory;
 import org.hibernate.search.backend.lucene.search.common.impl.LuceneSearchIndexScope;
 import org.hibernate.search.backend.lucene.search.common.impl.LuceneSearchIndexValueFieldContext;
 import org.hibernate.search.backend.lucene.types.codec.impl.AbstractLuceneNumericFieldCodec;
@@ -94,32 +93,13 @@ public abstract class AbstractLuceneMetricNumericFieldAggregation<F, E extends N
 		}
 	}
 
-	public static class Factory<F>
-			extends AbstractLuceneCodecAwareSearchQueryElementFactory<FieldMetricAggregationBuilder.TypeSelector,
-					F,
-					AbstractLuceneNumericFieldCodec<F, ?>> {
-
-		private final String operation;
-
-		protected Factory(AbstractLuceneNumericFieldCodec<F, ?> codec, String operation) {
-			super( codec );
-			this.operation = operation;
-		}
-
-		@Override
-		public FieldMetricAggregationBuilder.TypeSelector create(LuceneSearchIndexScope<?> scope,
-				LuceneSearchIndexValueFieldContext<F> field) {
-			return new TypeSelector<>( codec, scope, field, operation );
-		}
-	}
-
-	private static class TypeSelector<F> implements FieldMetricAggregationBuilder.TypeSelector {
+	protected static class TypeSelector<F> implements FieldMetricAggregationBuilder.TypeSelector {
 		private final AbstractLuceneNumericFieldCodec<F, ?> codec;
 		private final LuceneSearchIndexScope<?> scope;
 		private final LuceneSearchIndexValueFieldContext<F> field;
 		private final String operation;
 
-		private TypeSelector(AbstractLuceneNumericFieldCodec<F, ?> codec,
+		protected TypeSelector(AbstractLuceneNumericFieldCodec<F, ?> codec,
 				LuceneSearchIndexScope<?> scope, LuceneSearchIndexValueFieldContext<F> field,
 				String operation) {
 			this.codec = codec;
