+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/types/aggregation/impl/LuceneAvgNumericFieldAggregation.java
@@ -8,16 +8,18 @@ import org.hibernate.search.backend.lucene.lowlevel.aggregation.collector.impl.C
 import org.hibernate.search.backend.lucene.lowlevel.aggregation.collector.impl.SumCollectorFactory;
 import org.hibernate.search.backend.lucene.lowlevel.docvalues.impl.JoiningLongMultiValuesSource;
 import org.hibernate.search.backend.lucene.search.aggregation.impl.AggregationRequestContext;
+import org.hibernate.search.backend.lucene.search.common.impl.AbstractLuceneCodecAwareSearchQueryElementFactory;
 import org.hibernate.search.backend.lucene.search.common.impl.LuceneSearchIndexScope;
 import org.hibernate.search.backend.lucene.search.common.impl.LuceneSearchIndexValueFieldContext;
 import org.hibernate.search.backend.lucene.types.codec.impl.AbstractLuceneNumericFieldCodec;
 import org.hibernate.search.engine.backend.types.converter.spi.ProjectionConverter;
+import org.hibernate.search.engine.search.aggregation.spi.FieldMetricAggregationBuilder;
 
 public class LuceneAvgNumericFieldAggregation<F, E extends Number, K>
 		extends AbstractLuceneMetricNumericFieldAggregation<F, E, K> {
 
 	public static <F> Factory<F> factory(AbstractLuceneNumericFieldCodec<F, ?> codec) {
-		return new Factory<>( codec, "avg" );
+		return new Factory<>( codec );
 	}
 
 	LuceneAvgNumericFieldAggregation(Builder<F, E, K> builder) {
@@ -32,6 +34,22 @@ public class LuceneAvgNumericFieldAggregation<F, E extends Number, K>
 		countCollectorKey = countCollectorFactory.getCollectorKey();
 		context.requireCollector( sumCollectorFactory );
 		context.requireCollector( countCollectorFactory );
+	}
+
+	public static class Factory<F>
+			extends AbstractLuceneCodecAwareSearchQueryElementFactory<FieldMetricAggregationBuilder.TypeSelector,
+					F,
+					AbstractLuceneNumericFieldCodec<F, ?>> {
+
+		protected Factory(AbstractLuceneNumericFieldCodec<F, ?> codec) {
+			super( codec );
+		}
+
+		@Override
+		public FieldMetricAggregationBuilder.TypeSelector create(LuceneSearchIndexScope<?> scope,
+				LuceneSearchIndexValueFieldContext<F> field) {
+			return new TypeSelector<>( codec, scope, field, "avg" );
+		}
 	}
 
 	protected static class Builder<F, E extends Number, K>
