+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/types/aggregation/impl/LuceneMaxNumericFieldAggregation.java
@@ -7,16 +7,18 @@ package org.hibernate.search.backend.lucene.types.aggregation.impl;
 import org.hibernate.search.backend.lucene.lowlevel.aggregation.collector.impl.MaxCollectorFactory;
 import org.hibernate.search.backend.lucene.lowlevel.docvalues.impl.JoiningLongMultiValuesSource;
 import org.hibernate.search.backend.lucene.search.aggregation.impl.AggregationRequestContext;
+import org.hibernate.search.backend.lucene.search.common.impl.AbstractLuceneCodecAwareSearchQueryElementFactory;
 import org.hibernate.search.backend.lucene.search.common.impl.LuceneSearchIndexScope;
 import org.hibernate.search.backend.lucene.search.common.impl.LuceneSearchIndexValueFieldContext;
 import org.hibernate.search.backend.lucene.types.codec.impl.AbstractLuceneNumericFieldCodec;
 import org.hibernate.search.engine.backend.types.converter.spi.ProjectionConverter;
+import org.hibernate.search.engine.search.aggregation.spi.FieldMetricAggregationBuilder;
 
 public class LuceneMaxNumericFieldAggregation<F, E extends Number, K>
 		extends AbstractLuceneMetricNumericFieldAggregation<F, E, K> {
 
 	public static <F> Factory<F> factory(AbstractLuceneNumericFieldCodec<F, ?> codec) {
-		return new Factory<>( codec, "max" );
+		return new Factory<>( codec );
 	}
 
 	LuceneMaxNumericFieldAggregation(Builder<F, E, K> builder) {
@@ -28,6 +30,22 @@ public class LuceneMaxNumericFieldAggregation<F, E extends Number, K>
 		MaxCollectorFactory collectorFactory = new MaxCollectorFactory( source );
 		collectorKey = collectorFactory.getCollectorKey();
 		context.requireCollector( collectorFactory );
+	}
+
+	public static class Factory<F>
+			extends AbstractLuceneCodecAwareSearchQueryElementFactory<FieldMetricAggregationBuilder.TypeSelector,
+					F,
+					AbstractLuceneNumericFieldCodec<F, ?>> {
+
+		protected Factory(AbstractLuceneNumericFieldCodec<F, ?> codec) {
+			super( codec );
+		}
+
+		@Override
+		public FieldMetricAggregationBuilder.TypeSelector create(LuceneSearchIndexScope<?> scope,
+				LuceneSearchIndexValueFieldContext<F> field) {
+			return new TypeSelector<>( codec, scope, field, "max" );
+		}
 	}
 
 	protected static class Builder<F, E extends Number, K>
