+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/types/aggregation/impl/LuceneSumCompensatedSumAggregation.java
@@ -8,17 +8,19 @@ import org.hibernate.search.backend.lucene.lowlevel.aggregation.collector.impl.C
 import org.hibernate.search.backend.lucene.lowlevel.docvalues.impl.JoiningLongMultiValuesSource;
 import org.hibernate.search.backend.lucene.search.aggregation.impl.AggregationExtractContext;
 import org.hibernate.search.backend.lucene.search.aggregation.impl.AggregationRequestContext;
+import org.hibernate.search.backend.lucene.search.common.impl.AbstractLuceneCodecAwareSearchQueryElementFactory;
 import org.hibernate.search.backend.lucene.search.common.impl.LuceneSearchIndexScope;
 import org.hibernate.search.backend.lucene.search.common.impl.LuceneSearchIndexValueFieldContext;
 import org.hibernate.search.backend.lucene.types.codec.impl.AbstractLuceneNumericFieldCodec;
 import org.hibernate.search.backend.lucene.types.lowlevel.impl.LuceneNumericDomain;
 import org.hibernate.search.engine.backend.types.converter.spi.ProjectionConverter;
+import org.hibernate.search.engine.search.aggregation.spi.FieldMetricAggregationBuilder;
 
 public class LuceneSumCompensatedSumAggregation<F, E extends Number, K>
 		extends AbstractLuceneMetricCompensatedSumAggregation<F, E, K> {
 
 	public static <F> Factory<F> factory(AbstractLuceneNumericFieldCodec<F, ?> codec) {
-		return new Factory<>( codec, "sum" );
+		return new Factory<>( codec );
 	}
 
 	LuceneSumCompensatedSumAggregation(Builder<F, E, K> builder) {
@@ -38,6 +40,22 @@ public class LuceneSumCompensatedSumAggregation<F, E extends Number, K>
 	E extractEncoded(AggregationExtractContext context, LuceneNumericDomain<E> numericDomain) {
 		Double sum = context.getFacets( compensatedSumCollectorKey );
 		return numericDomain.doubleToTerm( sum );
+	}
+
+	public static class Factory<F>
+			extends AbstractLuceneCodecAwareSearchQueryElementFactory<FieldMetricAggregationBuilder.TypeSelector,
+					F,
+					AbstractLuceneNumericFieldCodec<F, ?>> {
+
+		protected Factory(AbstractLuceneNumericFieldCodec<F, ?> codec) {
+			super( codec );
+		}
+
+		@Override
+		public FieldMetricAggregationBuilder.TypeSelector create(LuceneSearchIndexScope<?> scope,
+				LuceneSearchIndexValueFieldContext<F> field) {
+			return new TypeSelector<>( codec, scope, field, "sum" );
+		}
 	}
 
 	protected static class Builder<F, E extends Number, K>
